backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
firebase-admin==6.4.0
python-dotenv==1.0.0
pydantic==2.5.0
numpy==1.26.2
requests==2.31.0

